minor(all): fixed minor bugs and added constants names and dictionary
